requirements.txt
# Core Dependencies
openai-whisper==20231117
anthropic==0.18.1
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Video Processing
ffmpeg-python==0.2.0
moviepy==1.0.3
pillow==10.2.0

# Data Validation
pydantic==2.5.3
validators==0.22.0

# Utilities
python-dateutil==2.8.2
colorama==0.4.6

# Optional: GPU Support for Whisper
# torch==2.1.2
# torchaudio==2.1.2